+++ CHAPTER6/main/DB.py
@@ -5,7 +5,7 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃ゲームに関する情報 
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-シーン      = None  # タイトル→プレイ→終了→タイトル...と遷移する
+シーン      = None  # タイトル画面→プレイ画面→終了画面→タイトル画面...と遷移する
 ボスシーン  = None  # None→ボス待ち→ボス対戦中→None...と遷移する
 プレイ時間  = 0     # 難易度の変化，出現タイミングの物差し
 難易度      = 0     # 敵の強さ、出現頻度の物差し
@@ -19,15 +19,15 @@
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
                     # 生成タイミング        # 消滅タイミング
 obj背景     = None  # main                  # なし
-obj特殊効果 = None  # シーン-タイトル       # なし
-obj自機共通 = None  # シーン-タイトル       # なし
-obj自機     = []    # シーン-タイトル       # シーン-プレイ
+obj特殊効果 = None  # シーン-タイトル画面   # なし
+obj自機共通 = None  # シーン-タイトル画面   # なし
+obj自機     = []    # シーン-タイトル画面   # シーン-プレイ画面
 obj敵機     = []    # 更新-出現             # 敵機-FN移動,FN衝突
 objアイテム = []    # 更新-出現             # 敵機-FN移動,FN衝突
 obj弾_自機  = []    # 更新-発射             # 弾-FN移動,FN衝突
 obj弾_敵機  = []    # 更新-発射             # 弾-FN移動,FN衝突
 obj爆発     = []    # 自機,敵機.弾-衝突     # 爆発-移動
-                    # シーン-プレイ
+                    # シーン-プレイ画面
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃インスタンス(処理用)
