+++ CHAPTER6/main/FN更新4_出現.py
@@ -31,7 +31,7 @@ class 出現クラス:
     def 実行(self):
 		#┬
         #○シーンを確認する
-        if main.DB.シーン != シーンID.プレイ:return
+        if main.DB.シーン != シーンID.プレイ画面:return
         if main.DB.プレイ時間 < 60 :return
         #│＼（対象外の場合）
         #│ ↓
@@ -56,7 +56,7 @@ class 出現クラス:
         #│ ↓
         #│ ▼処理を中断する
         #│
-        #○ボスシーンを『ボス対戦』にセットする
+        #○ボスシーンを『ボス対戦』に進行する
         main.DB.ボスシーン = シーンID.ボス対決
         #│
         #●特殊効果を解除(永続は残る)
