+++ CHAPTER6/main/main.py
@@ -27,8 +27,8 @@ class ゲーム本体:
         #〇処理で用いる各種オブジェクトを作成する
         self.初期化_処理セット()
         #│
-        #≫シーンを切替える(タイトル)
-        main.DB.FNシーン[ シーンID.終了 ].FN移動.Fn切替()
+        #≫シーンを『終了画面』に進行する
+        main.DB.FNシーン[ シーンID.終了画面 ].FN移動.Fn切替()
         #│
         #○フレーム処理に、処理セットを登録する
         pyxel.run(self.更新処理,self.描画処理)
@@ -70,9 +70,9 @@ class ゲーム本体:
         #│
         #≫シーン処理のオブジェクト・セットを作る(辞書型）
         main.DB.FNシーン = {
-            シーンID.タイトル   : タイトルクラス(),
-            シーンID.プレイ     : プレイクラス  (),
-            シーンID.終了       : 終了クラス    ()}
+            シーンID.タイトル画面   : タイトル画面(),
+            シーンID.プレイ画面     : プレイ画面()  ,
+            シーンID.終了画面       : 終了画面()    }
         #│
         #≫背景をオブジェクト化する
         main.DB.obj背景 = 背景作成()
