+++ CHAPTER6/シーン/DB.py
@@ -6,12 +6,10 @@ from enum import IntEnum, auto
 class シーンID(IntEnum):
 	#┬
 	#□シーン
-	タイトル		= 100
-	プレイ			= 200
-	ステージコール	= 201
+	タイトル画面	= 100
+	プレイ画面		= 200
 	ボス警告		= 300
 	ボス登場		= 301
 	ボス対決		= 302
-	ボーナス		= 400
-	終了			= 900
+	終了画面		= 900
 	#┴
