--- CHAPTER6/シーン/FNタイトル.py
@@ -1,127 +1,0 @@
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃シーン：タイトル
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-import pyxel
-import main.DB
-from   .DB             import シーンID
-from   キャラクタ.自機 import 自機登場, 自機共通生成
-from   特殊効果	       import 特殊効果作成
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃仕様
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 仕様クラス:
-    #--------------------------------------------------------------------
-    # 識別
-    #--------------------------------------------------------------------
-    #□種類ID    ：担当するシーン
-    種類ID = シーンID.タイトル
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃メイン
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 本体:
-    #┌───────────────────────────────────
-    #│初期化
-    #└───────────────────────────────────
-    def __init__(self):
-		#┬
-        #≫データセットを用意する
-        self.仕様   = 仕様クラス()
-        self.情報   = None
-        #│
-        #≫処理セットを用意する
-        self.FN移動 = 移動クラス(self)
-        self.FN描画 = 描画クラス(self)
-        #┴
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃移動機能
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 移動クラス:
-    #┌───────────────────────────────────
-    #│初期化
-    #└───────────────────────────────────
-    def __init__(self, 引数_生成先 ): #① 生成先のオブジェクト
-		#┬
-        #〇上位オブジェクトのリファレンスを用意する
-        self.基底 = 引数_生成先
-        #┴
-
-    #┌───────────────────────────────────
-    #│機能実行
-    #└───────────────────────────────────
-    def 実行(self):
-		#┬
-        #○シーン進行指示を確認する
-        人数確認 = 0
-        if pyxel.btn(pyxel.KEY_1): 人数確認 = 1
-        if pyxel.btn(pyxel.KEY_2): 人数確認 = 2
-        if 人数確認 == 0: return
-        #│＼（ユーザ操作の指示がされていない場合）
-        #│ ↓
-        #│ ▼処理を中断する
-        #│
-        #○シーンを『プレイ』に進行する
-        #○ボスシーンを初期化する
-        main.DB.シーン      = シーンID.プレイ
-        main.DB.ボスシーン  = None
-        #│
-        #○ゲーム情報をリセットする
-        main.DB.得点        = 0
-        main.DB.プレイ時間  = 1
-        main.DB.難易度      = 1
-        #│
-        #●自機共通の情報オブジェクトを生成する
-        #●特殊効果オブジェクトを生成する ※自機共通を参照するの為この順序
-        main.DB.obj自機共通 = 自機共通生成()
-        main.DB.obj特殊効果 = 特殊効果作成()        
-        #│
-        #●自機を生成する
-        main.DB.人数 = 人数確認
-        位置間隔 = int(pyxel.width / (main.DB.人数 + 1))
-        for tmpNo in range(main.DB.人数):
-            自機登場( tmpNo, 位置間隔 * (tmpNo + 1), 140)
-        #│
-        #○プレイ用のBGMを鳴らす
-        pyxel.stop()
-        pyxel.playm(1, loop=True)
-        #┴
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃描画機能
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 描画クラス:
-    #┌───────────────────────────────────
-    #│初期化
-    #└───────────────────────────────────
-    def __init__(self, 引数_生成先 ): #① 生成先のオブジェクト
-		#┬
-        #〇上位オブジェクトのリファレンスを用意する
-        self.基底 = 引数_生成先
-        #┴
-
-    #┌───────────────────────────────────
-    #│機能実行
-    #└───────────────────────────────────
-    def 実行(self):
-		#┬
-        #〇画面を描画する
-        pyxel.blt(0, 0, 2, 0, 0, 120, 120, 15)
-
-        描画文字 = "How Many Players? "
-        pyxel.text(27, 106, 描画文字, 5)
-        pyxel.text(26, 105, 描画文字, 7)
-
-        描画文字 = "Hit Any"
-        pyxel.text(11, 125, 描画文字, 5)
-        pyxel.text(10, 124, 描画文字, 7)
-
-        描画文字 = "Key    [1] or [2]"
-        pyxel.text(46, 121, 描画文字, 8)
-        pyxel.text(45, 120, 描画文字, 7)
-
-        描画文字 = "Button [A] or [B]"
-        pyxel.text(46, 129, 描画文字, 8)
-        pyxel.text(45, 128, 描画文字, 7)
-        #┴
