+++ CHAPTER6/シーン/プレイ画面/FN移動.py
@@ -1,51 +1,22 @@
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃シーン：プレイ
+#┃シーン：プレイ：移動機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import pyxel
 import main.DB
-from   .DB             import シーンID
-from   特殊効果        import 効果ID
+from   ..DB            import シーンID
 from   キャラクタ.爆発 import 爆発開始
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃仕様
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 仕様クラス:
-    #--------------------------------------------------------------------
-    # 識別
-    #--------------------------------------------------------------------
-    #□種類ID    ：担当するシーン
-    種類ID = シーンID.プレイ
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃メイン
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 本体:
-    #┌───────────────────────────────────
-    #│初期化
-    #└───────────────────────────────────
-    def __init__(self):
-		#┬
-        #≫データセットを用意する
-        self.仕様   = 仕様クラス()
-        self.情報   = None
-        #│
-        #≫処理セットを用意する
-        self.FN移動 = 移動クラス(self)
-        self.FN描画 = 描画クラス(self)
-        #┴
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃移動機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 移動_仕様クラス:
+class 仕様クラス:
     #--------------------------------------------------------------------
     # カウンタ
     #--------------------------------------------------------------------
     #□ボス警告時間      ：
     ボス警告時間        = 200
 
-class 移動_情報クラス:
+class 情報クラス:
     #--------------------------------------------------------------------
     # カウンタ
     #--------------------------------------------------------------------
@@ -56,12 +27,11 @@ class 移動クラス:
     #┌───────────────────────────────────
     #│初期化
     #└───────────────────────────────────
-    def __init__(self,
-            引数_生成先 ):  #① 生成先のオブジェクト
+    def __init__(self):
 		#┬
         #≫データセットを用意する
-        self.仕様 = 移動_仕様クラス()
-        self.情報 = 移動_情報クラス()
+        self.仕様 = 仕様クラス()
+        self.情報 = 情報クラス()
         #┴
 
     #┌───────────────────────────────────
@@ -91,8 +61,8 @@ class 移動クラス:
         #│ ↓
         #│ ▼処理を中断する
         #│
-        #○シーンを『終了』に進行する
-        main.DB.シーン = シーンID.終了
+        #○シーンを『終了画面』に進行する
+        main.DB.シーン = シーンID.終了画面
         #│
         #◎└┐自機をすべて爆破する
         所有者 = main.DB.所有者ID.自機
@@ -124,7 +94,7 @@ class 移動クラス:
 	#────────────────────────────────────
     def Fnプレイ進行(self):
 		#┬
-        #◇┐状況に応じてプレイシーンを進行する 
+        #◇┐状況に応じてプレイ画面を進行する 
         if main.DB.ボスシーン == シーンID.ボス警告:
         #　├┐（ボス襲来で警告中の場合）
             #↓
@@ -166,65 +136,4 @@ class 移動クラス:
             #○難易度をセットする
             main.DB.プレイ時間 += 1
             main.DB.難易度 = main.DB.プレイ時間 // main.DB.難易度間隔 + 1
-        #┴　┴
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃描画機能
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 描画クラス:
-    #┌───────────────────────────────────
-    #│初期化
-    #└───────────────────────────────────
-    def __init__(self, 引数_生成先 ): #① 生成先のオブジェクト
-		#┬
-        #〇上位オブジェクトのリファレンスを用意する
-        self.基底 = 引数_生成先
-        #┴
-
-    #┌───────────────────────────────────
-    #│機能実行
-    #└───────────────────────────────────
-    def 実行(self):
-		#┬
-        #〇リファレンスを用意する
-        if not main.DB.obj自機: return
-        #│＼（自機が存在しない場合）
-        #│ ↓
-        #│ ▼処理を中断する
-        #│
-        #○効果発動の登録状況を用意する
-        発動中 = main.DB.obj特殊効果.情報.発動中
-        #│
-        共通情報 = main.DB.obj自機共通.情報
-        画面縦長 = pyxel.height
-        #│
-        #○残りのシールドを描画する
-        数量 = int(共通情報.シールド)
-        pyxel.rect(0, 画面縦長 - 12, 数量, 3, 7)
-        #│
-        #◎└┐残りの弾薬を描画する
-        座標X   = 1
-        座標Y   = 画面縦長 - 8
-        画像X   = (4) if 効果ID.貫通弾 in 発動中 else (0)
-        数量 = int(共通情報.弾数)
-        for i in range(数量):
-			#│＼（すべての処理を終えた場合）
-            #│ ↓
-			#│ ▼繰り返し処理を抜ける
-            #│
-            #○ひとつずつ描画する
-            pyxel.blt(座標X, 座標Y, 0, 画像X, 40, 4, 8, 0)
-            座標X += 5
-            #┴
-        #│
-        #◇┐ボス襲来の警告を表示する。 
-        if main.DB.ボスシーン == シーンID.ボス警告 :
-        #　├┐（ボス待ちの場合）
-            #↓
-            #○残りのシールドを描画する
-            メッセージ = "Enemy aircraft encounter"
-            pyxel.text(16, 41, メッセージ, 0)
-            pyxel.text(15, 40, メッセージ, pyxel.rndi(1,15))
-            #┴
-        #　└┐（その他）
         #┴　┴
