+++ CHAPTER6/シーン/プレイ画面/main.py
@@ -1,0 +1,19 @@
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃シーン：プレイ
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+from   .FN移動 import 移動クラス
+from   .FN描画 import 描画クラス
+
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃メイン
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+class 本体:
+    #┌───────────────────────────────────
+    #│初期化
+    #└───────────────────────────────────
+    def __init__(self):
+		#┬
+        #≫処理セットを用意する
+        self.FN移動 = 移動クラス()
+        self.FN描画 = 描画クラス()
+        #┴
