+++ CHAPTER6/シーン/終了画面/FN描画.py
@@ -1,0 +1,25 @@
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃シーン：終了：描画機能
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+import pyxel
+
+#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+#┃描画機能
+#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
+class 描画クラス:
+    #┌───────────────────────────────────
+    #│初期化
+    #└───────────────────────────────────
+    def __init__(self):
+      pass
+
+    #┌───────────────────────────────────
+    #│機能実行
+    #└───────────────────────────────────
+    def 実行(self):
+      #┬
+      #〇画面を描画する
+      描画文字 = "- GAME OVER -"
+      pyxel.text(36, 79, 描画文字, 5)
+      pyxel.text(35, 78, 描画文字, pyxel.frame_count % 16)
+      #┴
