+++ CHAPTER6/シーン/終了画面/FN移動.py
@@ -3,18 +3,12 @@
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 import pyxel
 import main.DB
-from   .DB import シーンID
+from   ..DB import シーンID
 
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃仕様
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 class 仕様クラス:
-    #--------------------------------------------------------------------
-    # 識別
-    #--------------------------------------------------------------------
-    #□種類ID    ：担当するシーン
-    種類ID          = シーンID.終了
-
     #--------------------------------------------------------------------
     # カウンタ
     #--------------------------------------------------------------------
@@ -33,24 +27,6 @@ class 情報クラス:
         self.待ち開始       = True
         self.開始カウント   = 0
 
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃メイン
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 本体:
-    #┌───────────────────────────────────
-    #│初期化
-    #└───────────────────────────────────
-    def __init__(self):
-		#┬
-        #≫データセットを用意する
-        self.仕様   = 仕様クラス()
-        self.情報   = 情報クラス()
-        #│
-        #≫処理セットを用意する
-        self.FN移動 = 移動クラス(self)
-        self.FN描画 = 描画クラス(self)
-        #┴
-
 #┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 #┃移動機能
 #┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
@@ -58,10 +34,11 @@ class 移動クラス:
     #┌───────────────────────────────────
     #│初期化
     #└───────────────────────────────────
-    def __init__(self, 引数_生成先 ): #① 生成先のオブジェクト
+    def __init__(self):
 		#┬
-        #〇上位オブジェクトのリファレンスを用意する
-        self.基底 = 引数_生成先
+        #〇個体オブジェクトのリファレンスを用意する
+        self.仕様  = 仕様クラス()
+        self.情報  = 情報クラス()
         #┴
 
     #┌───────────────────────────────────
@@ -69,10 +46,6 @@ class 移動クラス:
     #└───────────────────────────────────
     def 実行(self):
 		#┬
-        #○上位オブジェクトのリファレンスを用意する
-        仕様 = self.基底.仕様
-        情報 = self.基底.情報
-        #│
         #○爆発状況を確認する
         if  main.DB.obj爆発: return
         #│＼（爆発中の場合）
@@ -80,7 +53,7 @@ class 移動クラス:
         #│ ▼処理を中断する
         #│
         #◇┐ＢＧＭを開始する
-        if 情報.待ち開始:
+        if self.情報.待ち開始:
         #　├┐（待ちを開始する場合）
             #↓
             #○終了用のBGMを鳴らす
@@ -88,15 +61,15 @@ class 移動クラス:
             pyxel.playm(7, loop=True)
             #│
             #○待ち開始をオフする
-            情報.待ち開始       = False
-            情報.開始カウント   = pyxel.frame_count
+            self.情報.待ち開始       = False
+            self.情報.開始カウント   = pyxel.frame_count
             #┴
         #　└┐（その他）
             #┴
         #│
-        #◇┐次のシーンに切替える
-        経過カウント = pyxel.frame_count - 情報.開始カウント
-        if (経過カウント > 仕様.待ちカウント):
+        #◇┐次のシーンに進行する
+        経過カウント = pyxel.frame_count - self.情報.開始カウント
+        if (経過カウント > self.仕様.待ちカウント):
         #　├┐（待ちを時間を経過した場合）
             #↓
             #●シーンを切替える
@@ -106,11 +79,8 @@ class 移動クラス:
 	#────────────────────────────────────	
     def Fn切替(self):
 		#┬
-        #○上位オブジェクトのリファレンスを用意する
-        情報 = self.基底.情報
-        #│
-        #○シーンを『タイトル』に進行する
-        main.DB.シーン = シーンID.タイトル
+        #○シーンを『タイトル画面』に進行する
+        main.DB.シーン = シーンID.タイトル画面
         #│
         #○└┐インスタンスを初期化する
             #●敵機を抹消する
@@ -124,32 +94,8 @@ class 移動クラス:
             #┴
         #│
         #○待ち開始をオンにする
-        情報.待ち開始 = True
+        self.情報.待ち開始 = True
         #│
         #○タイトル用のBGMを鳴らす
         pyxel.playm(0, loop=True)
-        #┴
-
-#┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-#┃描画機能
-#┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
-class 描画クラス:
-    #┌───────────────────────────────────
-    #│初期化
-    #└───────────────────────────────────
-    def __init__(self, 引数_生成先 ): #① 生成先のオブジェクト
-		#┬
-        #〇上位オブジェクトのリファレンスを用意する
-        self.基底 = 引数_生成先
-        #┴
-
-    #┌───────────────────────────────────
-    #│機能実行
-    #└───────────────────────────────────
-    def 実行(self):
-		#┬
-        #〇画面を描画する
-        描画文字 = "- GAME OVER -"
-        pyxel.text(36, 79, 描画文字, 5)
-        pyxel.text(35, 78, 描画文字, pyxel.frame_count % 16)
         #┴
