+++ CHAPTER6/背景/FN描画.py
@@ -27,7 +27,7 @@ class 描画クラス:
         情報 = self.基底.情報
         #│
         #◇┐銀河を描画する
-        if main.DB.シーン != シーンID.タイトル:
+        if main.DB.シーン != シーンID.タイトル画面:
         #　├┐（シーンが『タイトル』以外の場合）
             #↓
             #○銀河を描画する
